Add tests for quote generator
Huge thanks to @mrbenbot for the super clean tests to copy from in #67!

This adds a few requirements to the problem specification so that
specific IDs must be used.

It also adds an implicit requirement in the tests that the "new quote"
button should always change the quote (which isn't explicitly made clear
in the description).